refactor(transfer): split receiver transfer setup into submodules
